+++ src/AppStream.DurablePatterns/Executor/ExecutionResult.cs
@@ -1,10 +1,5 @@
-﻿using AppStream.DurablePatterns.Builder;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-
-namespace AppStream.DurablePatterns
+﻿namespace AppStream.DurablePatterns
 {
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
     public class ExecutionResult
     {
         public ExecutionResult(IEnumerable<StepExecutionResultSummary> steps)
@@ -15,12 +10,10 @@ namespace AppStream.DurablePatterns
         public IEnumerable<StepExecutionResultSummary> Steps { get; }
     }
 
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
     public record StepExecutionResultSummary(
         string Name,
         TimeSpan Duration);
 
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
     public record FanOutFanInStepExecutionResultSummary(
         string Name,
         TimeSpan Duration,