+++ src/AppStream.DurablePatterns/Executor/IDurablePatternsExecutor.cs
@@ -1,13 +1,12 @@
 ﻿using AppStream.DurablePatterns.Steps;
-using Microsoft.Azure.WebJobs.Extensions.DurableTask;
+using Microsoft.DurableTask;
 
 namespace AppStream.DurablePatterns.Executor
 {
     internal interface IDurablePatternsExecutor
     {
         Task<ExecutionResult> ExecuteAsync(
-            IEnumerable<Step> steps, 
-            EntityId stepsEntityId,
-            IDurableOrchestrationContext context);
+            IEnumerable<Step> steps,
+            TaskOrchestrationContext context);
     }
 }