+++ src/AppStream.DurablePatterns/Steps/Entity/IStepsEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace AppStream.DurablePatterns.Steps.Entity
-{
-    public interface IStepsEntity
-    {
-        Task Set(Dictionary<Guid, Step> steps);
-    }
-}