+++ test/AppStream.DurablePatterns.Tests/FanOutFanInStepExecutorTests.cs
@@ -2,10 +2,9 @@
 using AppStream.DurablePatterns.Executor.StepExecutor.FanOutFanInStep;
 using AppStream.DurablePatterns.Executor.StepExecutor.FanOutFanInStep.OptionsValidator;
 using AppStream.DurablePatterns.Steps;
-using AppStream.DurablePatterns.Steps.Entity;
-using Microsoft.Azure.WebJobs.Extensions.DurableTask;
+using Microsoft.DurableTask;
 using Moq;
-using Newtonsoft.Json.Linq;
+using System.Text.Json;
 
 namespace AppStream.DurablePatterns.Tests
 {
@@ -26,127 +25,126 @@ namespace AppStream.DurablePatterns.Tests
         public void ExecuteStepAsync_InputIsNull_ThrowsArgumentNullException()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(List<string>),
-                typeof(List<string>),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
                 new FanOutFanInOptions(1, 1));
 
-            var context = Mock.Of<IDurableOrchestrationContext>();
+            var context = Mock.Of<TaskOrchestrationContext>();
 
             // Act & Assert
             Assert.ThrowsAsync<ArgumentNullException>(
-                () => _executor.ExecuteStepAsync(step, stepsEntityId, context, null));
+                () => _executor.ExecuteStepAsync(step, context, null));
         }
 
         [Test]
         public void ExecuteStepAsync_InputIsNotCollection_ThrowsArgumentException()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(string),
-                typeof(List<string>),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(string).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
                 new FanOutFanInOptions(1, 1));
 
-            var context = Mock.Of<IDurableOrchestrationContext>();
+            var context = Mock.Of<TaskOrchestrationContext>();
             var input = "not a collection";
 
             // Act & Assert
             Assert.ThrowsAsync<ArgumentException>(
-                () => _executor.ExecuteStepAsync(step, stepsEntityId, context, input));
+                () => _executor.ExecuteStepAsync(step, context, input));
         }
 
         [Test]
         public void ExecuteStepAsync_ResultTypeIsNotCollection_ThrowsArgumentException()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(List<string>),
-                typeof(string),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
+                typeof(string).AssemblyQualifiedName!,
                 new FanOutFanInOptions(1, 1));
 
-            var context = Mock.Of<IDurableOrchestrationContext>();
+            var context = Mock.Of<TaskOrchestrationContext>();
             var input = new List<string>();
 
             // Act & Assert
             Assert.ThrowsAsync<ArgumentException>(
-                () => _executor.ExecuteStepAsync(step, stepsEntityId, context, input));
+                () => _executor.ExecuteStepAsync(step, context, input));
         }
 
         [Test]
         public async Task ExecuteFanOutFanInInternalAsync_InputsAreValid_CallsActivityFunction()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(List<string>),
-                typeof(List<string>),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
                 new FanOutFanInOptions(3, 1));
             var input = new List<string>() { "1", "2", "3" };
 
-            var contextMock = new Mock<IDurableOrchestrationContext>();
+            var contextMock = new Mock<TaskOrchestrationContext>();
             contextMock
                 .Setup(c => c.CallActivityAsync<ActivityFunctionResult>(
                     ActivityFunction.FunctionName,
-                    It.Is<ActivityFunctionInput>(i => i.StepId == step.StepId)))
+                    It.Is<ActivityFunctionInput>(i => i.Step.StepId == step.StepId),
+                    It.IsAny<TaskOptions?>()))
                 .ReturnsAsync(new ActivityFunctionResult(
-                    JToken.FromObject(new List<string>() { "1", "2", "3" }), 
+                    JsonSerializer.SerializeToElement(new List<string>() { "1", "2", "3" }),
                     TimeSpan.Zero));
 
             // Act
-            var result = await _executor.ExecuteStepAsync(step, stepsEntityId, contextMock.Object, input);
+            var result = await _executor.ExecuteStepAsync(step, contextMock.Object, input);
             
             // Assert
             contextMock.Verify(m => m.CallActivityAsync<ActivityFunctionResult>(
                 ActivityFunction.FunctionName, 
-                It.Is<ActivityFunctionInput>(i => i.StepId == step.StepId)));
+                It.Is<ActivityFunctionInput>(i => i.Step.StepId == step.StepId),
+                It.IsAny<TaskOptions?>()));
         }
 
         [Test]
         public async Task ExecuteFanOutFanInInternalAsync_InputsAreValid_CallsActivityFunctionOnceForEachBatch()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(List<string>),
-                typeof(List<string>),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName!,
                 new FanOutFanInOptions(2, 1));
             var input = new List<string>() { "1", "2", "3", "4", "5", "6", "7" };
 
-            var contextMock = new Mock<IDurableOrchestrationContext>();
+            var contextMock = new Mock<TaskOrchestrationContext>();
             contextMock
                 .Setup(c => c.CallActivityAsync<ActivityFunctionResult>(
                     ActivityFunction.FunctionName,
-                    It.Is<ActivityFunctionInput>(i => i.StepId == step.StepId)))
-                .ReturnsAsync((string functionName, ActivityFunctionInput functionInput) => new ActivityFunctionResult(
-                    JToken.FromObject(functionInput.Input!),
+                    It.Is<ActivityFunctionInput>(i => i.Step.StepId == step.StepId), 
+                    It.IsAny<TaskOptions?>()))
+                .ReturnsAsync((TaskName functionName, ActivityFunctionInput functionInput, TaskOptions options) => new ActivityFunctionResult(
+                    JsonSerializer.SerializeToElement(functionInput.ActivityInput!),
                     TimeSpan.Zero));
 
             // Act
-            var result = await _executor.ExecuteStepAsync(step, stepsEntityId, contextMock.Object, input);
+            var result = await _executor.ExecuteStepAsync(step, contextMock.Object, input);
 
             // Assert
             contextMock.Verify(
                 m => m.CallActivityAsync<ActivityFunctionResult>(
                     ActivityFunction.FunctionName,
-                    It.Is<ActivityFunctionInput>(i => i.StepId == step.StepId)), 
+                    It.Is<ActivityFunctionInput>(i => i.Step.StepId == step.StepId), 
+                    It.IsAny<TaskOptions?>()), 
                 Times.Exactly(4));
         }
 
@@ -154,28 +152,28 @@ namespace AppStream.DurablePatterns.Tests
         public async Task CombineResults_InputsAreValid_ReturnsCombinedResults()
         {
             // Arrange
-            var stepsEntityId = new EntityId(nameof(StepsEntity), Guid.NewGuid().ToString());
             var step = new Step(
                 Guid.NewGuid(),
                 StepType.FanOutFanIn,
-                typeof(MyPatternActivity),
-                typeof(List<string>),
-                typeof(List<string>),
+                typeof(MyPatternActivity).AssemblyQualifiedName!,
+                typeof(List<string>).AssemblyQualifiedName! ,
+                typeof(List<string>).AssemblyQualifiedName!,
                 new FanOutFanInOptions(2, 1));
             var input = new List<string>() { "1", "2", "3", "4", "5", "6", "7" };
             var expectedResult = new List<string>() { "1", "1", "2", "2", "3", "3", "4", "4", "5", "5", "6", "6", "7", "7" };
 
-            var contextMock = new Mock<IDurableOrchestrationContext>();
+            var contextMock = new Mock<TaskOrchestrationContext>();
             contextMock
                 .Setup(c => c.CallActivityAsync<ActivityFunctionResult>(
                     ActivityFunction.FunctionName,
-                    It.Is<ActivityFunctionInput>(i => i.StepId == step.StepId)))
-                .ReturnsAsync((string functionName, ActivityFunctionInput functionInput) => new ActivityFunctionResult(
-                    JToken.FromObject(((List<string>)functionInput.Input!).Concat((List<string>)functionInput.Input!)),
+                    It.Is<ActivityFunctionInput>(i => i.Step.StepId == step.StepId), 
+                    It.IsAny<TaskOptions?>()))
+                .ReturnsAsync((TaskName functionName, ActivityFunctionInput functionInput, TaskOptions options) => new ActivityFunctionResult(
+                    JsonSerializer.SerializeToElement(((List<string>)functionInput.ActivityInput!).Concat((List<string>)functionInput.ActivityInput!)),
                     TimeSpan.Zero));
 
             // Act
-            var result = await _executor.ExecuteStepAsync(step, stepsEntityId, contextMock.Object, input);
+            var result = await _executor.ExecuteStepAsync(step, contextMock.Object, input);
             
             // Assert
             Assert.Multiple(() =>